+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+- Add configuration to plugin which allow it to return validation error back
+  to the client as part of request response 
+  
 ## 0.4.0
 
 - bump lua-resty-jsonschema to fix issue with too many local variables